introduced bcrypt and signup method that will store the users information in the database and hash + salt the password first
