+++ backend/modules/auth.ts
@@ -1,0 +1,5 @@
+import bcrypt from 'bcrypt';
+
+export const getHashedPlusSaltedPassword = (password: string) => {
+  return bcrypt.hash(password, 5);
+};
